Refactor monolithic Firebase functions into modular architecture
- Split 1500-line index.js into 5 domain-specific modules (questions, quizzes, results, flashcards, study) with 17 functions total
- Improved developer onboarding experience with clear separation of concerns and faster cold starts
- Refactored quizService.js into specialized services (quizCreationService, quizRetrievalService, quizSubmissionService, analyticsService)
- Updated components and contexts to use new modular service layer
- Enhanced documentation and removed redundant files for cleaner codebase
- All functions deployed and tested successfully
