+++ functions/src/flashcards/index.js
@@ -1,0 +1,338 @@
+const functions = require('firebase-functions')
+const {onRequest, onCall} = require('firebase-functions/v2/https')
+const admin = require('firebase-admin')
+const cors = require("cors")({origin: true})
+
+/**
+ * Create a new flashcard deck
+ */
+exports.addCustomFlashcardDeck = onRequest(async (req, res) => {
+    cors(req, res, async () => {
+        const dataType = req.get('content-type');
+        if (dataType === 'application/json') {
+            const data = JSON.parse(JSON.stringify(req.body));
+
+            // Extract data from request
+            const creatorID = data.creator?.uid || data.creatorID;
+            const title = data.metadata?.title || data.name || data.title;
+            const cards = data.content?.cards || data.cards || [];
+            const tags = data.metadata?.tags || data.tags || "";
+
+            // Validation
+            if (!creatorID || !title.trim() || !Array.isArray(cards) || cards.length === 0) {
+                return res.json({
+                    status: 400,
+                    success: false,
+                    message: "Missing required parameters: creatorID, title, and cards array"
+                });
+            }
+
+            try {
+                const user = await admin.firestore().collection('users').doc(creatorID);
+                const userDoc = await user.get();
+
+                // Get creator information
+                let creatorInfo = {
+                    uid: creatorID,
+                    displayName: 'Anonymous User',
+                    username: 'Anonymous User'
+                };
+
+                if (userDoc.exists) {
+                    const userData = userDoc.data();
+                    creatorInfo = {
+                        uid: creatorID,
+                        displayName: userData.profile?.displayName || userData.displayName ||
+                                   `${userData.profile?.firstName || 'Anonymous'} ${userData.profile?.lastName || 'User'}`.trim(),
+                        username: userData.profile?.displayName || userData.displayName || 'Anonymous User'
+                    };
+                }
+
+                const currentDate = new Date().toISOString();
+
+                // Convert cards array to map format like custom_quizzes
+                const cardsMap = {};
+                cards.forEach((card, index) => {
+                    const cardKey = `Card ${index + 1}`;
+                    cardsMap[cardKey] = {
+                        front: card.front || '',
+                        back: card.back || '',
+                        type: card.type || 'basic'
+                    };
+                });
+
+                const newFlashcardDeck = {
+                    // Metadata section
+                    metadata: {
+                        title: title,
+                        description: data.metadata?.description || data.description || "",
+                        category: data.metadata?.category || data.category || "General",
+                        tags: Array.isArray(tags) ? tags.join(', ') : (tags || ""),
+                        cardCount: cards.length,
+                        isPublic: data.metadata?.isPublic || data.isPublic || false,
+                        difficulty: data.metadata?.difficulty || data.difficulty || "2",
+                        version: 1
+                    },
+
+                    // Creator information
+                    creator: creatorInfo,
+
+                    // Content - cards stored as map
+                    content: {
+                        cards: cardsMap
+                    },
+
+                    // Analytics - initialize empty
+                    analytics: {
+                        stats: {
+                            timesStudied: 0,
+                            averageScore: 0,
+                            lastStudied: null,
+                            totalReviews: 0
+                        },
+                        performance: {
+                            cardStats: {}
+                        }
+                    },
+
+                    // Access control
+                    access: {
+                        visibility: data.metadata?.isPublic || data.isPublic ? "public" : "private",
+                        allowCopying: data.access?.allowCopying || true,
+                        studyMode: data.access?.studyMode || "flashcards"
+                    },
+
+                    // Timestamps
+                    timestamps: {
+                        createdAt: currentDate,
+                        updatedAt: currentDate,
+                        lastStudiedAt: null
+                    },
+
+                    // Moderation
+                    moderation: {
+                        status: "active",
+                        reports: [],
+                        flags: []
+                    }
+                };
+
+                // Save to Firestore
+                const result = await admin.firestore().collection('flashcard_decks').add(newFlashcardDeck);
+
+                // Update user stats and cache (similar to quiz creation)
+                try {
+                    if (userDoc.exists) {
+                        await user.update({
+                            'stats.flashcardDecksCreated': admin.firestore.FieldValue.increment(1),
+                            'cache.recentFlashcardIds': admin.firestore.FieldValue.arrayUnion(result.id),
+                            'timestamps.updatedAt': currentDate,
+                            'timestamps.lastActiveAt': currentDate
+                        });
+                    }
+                } catch (error) {
+                    // Error updating user stats - silent fail
+                    console.error('Error updating user stats for flashcard deck:', error);
+                }
+
+                return res.json({
+                    status: 200,
+                    success: true,
+                    message: "Flashcard deck created successfully",
+                    deckID: result.id,
+                    data: newFlashcardDeck
+                });
+
+            } catch (error) {
+                console.error('Error creating flashcard deck:', error);
+                return res.json({
+                    status: 500,
+                    success: false,
+                    message: error.message
+                });
+            }
+        } else {
+            return res.json({
+                status: 400,
+                success: false,
+                message: "Content-Type must be application/json"
+            });
+        }
+    });
+});
+
+/**
+ * Get flashcard decks by user
+ */
+exports.getUserFlashcardDecks = onRequest(async (req, res) => {
+    cors(req, res, async () => {
+        const userId = req.query.userId || req.body?.userId;
+
+        if (!userId) {
+            return res.json({
+                status: 400,
+                success: false,
+                message: "User ID is required"
+            });
+        }
+
+        try {
+            const query = admin.firestore()
+                .collection('flashcard_decks')
+                .where('creator.uid', '==', userId)
+                .where('moderation.status', '==', 'active')
+                .orderBy('timestamps.updatedAt', 'desc');
+
+            const querySnapshot = await query.get();
+            const decks = [];
+
+            querySnapshot.forEach(doc => {
+                const data = doc.data();
+                decks.push({
+                    id: doc.id,
+                    ...data
+                });
+            });
+
+            return res.json({
+                status: 200,
+                success: true,
+                message: "Flashcard decks retrieved successfully",
+                data: decks,
+                count: decks.length
+            });
+
+        } catch (error) {
+            console.error('Error fetching user flashcard decks:', error);
+            return res.json({
+                status: 500,
+                success: false,
+                message: error.message
+            });
+        }
+    });
+});
+
+/**
+ * Get a specific flashcard deck by ID
+ */
+exports.getFlashcardDeck = onRequest(async (req, res) => {
+    cors(req, res, async () => {
+        const deckId = req.query.deckId || req.body?.deckId;
+
+        if (!deckId) {
+            return res.json({
+                status: 400,
+                success: false,
+                message: "Deck ID is required"
+            });
+        }
+
+        try {
+            const deckDoc = await admin.firestore().collection('flashcard_decks').doc(deckId).get();
+
+            if (!deckDoc.exists) {
+                return res.json({
+                    status: 404,
+                    success: false,
+                    message: "Flashcard deck not found"
+                });
+            }
+
+            const deckData = deckDoc.data();
+
+            return res.json({
+                status: 200,
+                success: true,
+                message: "Flashcard deck retrieved successfully",
+                data: {
+                    id: deckDoc.id,
+                    ...deckData
+                }
+            });
+
+        } catch (error) {
+            console.error('Error fetching flashcard deck:', error);
+            return res.json({
+                status: 500,
+                success: false,
+                message: error.message
+            });
+        }
+    });
+});
+
+/**
+ * Delete a flashcard deck
+ */
+exports.deleteFlashcardDeck = onRequest(async (req, res) => {
+    cors(req, res, async () => {
+        const deckId = req.query.deckId || req.body?.deckId;
+        const userId = req.query.userId || req.body?.userId;
+
+        if (!deckId || !userId) {
+            return res.json({
+                status: 400,
+                success: false,
+                message: "Deck ID and User ID are required"
+            });
+        }
+
+        try {
+            const deckDoc = await admin.firestore().collection('flashcard_decks').doc(deckId).get();
+
+            if (!deckDoc.exists) {
+                return res.json({
+                    status: 404,
+                    success: false,
+                    message: "Flashcard deck not found"
+                });
+            }
+
+            const deckData = deckDoc.data();
+
+            // Verify ownership
+            if (deckData.creator.uid !== userId) {
+                return res.json({
+                    status: 403,
+                    success: false,
+                    message: "Unauthorized: You can only delete your own decks"
+                });
+            }
+
+            // Soft delete by updating moderation status
+            await admin.firestore().collection('flashcard_decks').doc(deckId).update({
+                'moderation.status': 'deleted',
+                'timestamps.updatedAt': new Date().toISOString(),
+                'timestamps.deletedAt': new Date().toISOString()
+            });
+
+            // Update user stats: decrement deck count and remove from recent cache
+            try {
+                const user = admin.firestore().collection('users').doc(userId);
+                await user.update({
+                    'stats.flashcardDecksCreated': admin.firestore.FieldValue.increment(-1),
+                    'cache.recentFlashcardIds': admin.firestore.FieldValue.arrayRemove(deckId),
+                    'timestamps.updatedAt': new Date().toISOString()
+                });
+            } catch (error) {
+                // Error updating user stats - silent fail
+                console.error('Error updating user stats for flashcard deck deletion:', error);
+            }
+
+            return res.json({
+                status: 200,
+                success: true,
+                message: "Flashcard deck deleted successfully"
+            });
+
+        } catch (error) {
+            console.error('Error deleting flashcard deck:', error);
+            return res.json({
+                status: 500,
+                success: false,
+                message: error.message
+            });
+        }
+    });
+});
