+++ src/services/resultService.js
@@ -1,9 +1,8 @@
 /**
- * Result service - handles quiz results and analytics
+ * Result Service - handles basic quiz result retrieval
  */
 import { collection, doc, getDoc, getDocs, query, where, orderBy, limit, startAfter } from 'firebase/firestore';
-import { httpsCallable } from 'firebase/functions';
-import { db, functions, handleFirebaseError, withRetry, timestamp } from './firebaseService';
+import { db, handleFirebaseError, withRetry, timestamp } from './firebaseService';
 
 class ResultService {
     constructor() {
@@ -18,7 +17,7 @@ class ResultService {
      */
     async getAttemptById(attemptId) {
         try {
-            const attemptDoc = await withRetry(() => 
+            const attemptDoc = await withRetry(() =>
                 getDoc(doc(db, this.attemptsCollection, attemptId))
             );
 
@@ -68,11 +67,11 @@ class ResultService {
             if (startAfterDoc) {
                 q = query(q, startAfter(startAfterDoc));
             }
-            
+
             q = query(q, limit(limitCount));
 
             const querySnapshot = await getDocs(q);
-            
+
             const attempts = querySnapshot.docs.map(doc => ({
                 id: doc.id,
                 ...doc.data(),
@@ -113,11 +112,11 @@ class ResultService {
             if (startAfterDoc) {
                 q = query(q, startAfter(startAfterDoc));
             }
-            
+
             q = query(q, limit(limitCount));
 
             const querySnapshot = await getDocs(q);
-            
+
             const attempts = querySnapshot.docs.map(doc => ({
                 id: doc.id,
                 ...doc.data(),
@@ -137,100 +136,6 @@ class ResultService {
     }
 
     /**
-     * Get detailed analytics for a quiz
-     * @param {string} quizId - Quiz ID
-     * @returns {Promise<Object>} Quiz analytics
-     */
-    async getQuizAnalytics(quizId) {
-        try {
-            const analyticsFunction = httpsCallable(functions, 'getQuizAnalytics');
-            const result = await analyticsFunction({ quizId });
-        
-            return result.data;
-
-        } catch (error) {
-            throw handleFirebaseError(error);
-        }
-    }
-
-    /**
-     * Get user performance summary
-     * @param {string} userId - User ID
-     * @returns {Promise<Object>} User performance data
-     */
-    async getUserPerformanceSummary(userId) {
-        try {
-            const summaryFunction = httpsCallable(functions, 'getUserPerformanceSummary');
-            const result = await summaryFunction({ userId });
-        
-            return result.data;
-
-        } catch (error) {
-            throw handleFirebaseError(error);
-        }
-    }
-
-    /**
-     * Export quiz results to CSV format
-     * @param {string} quizId - Quiz ID
-     * @param {Object} options - Export options
-     * @returns {Promise<string>} CSV data
-     */
-    async exportQuizResults(quizId, options = {}) {
-        try {
-            const exportFunction = httpsCallable(functions, 'exportQuizResults');
-            const result = await exportFunction({ 
-                quizId, 
-                ...options 
-            });
-        
-            return result.data;
-
-        } catch (error) {
-            throw handleFirebaseError(error);
-        }
-    }
-
-    /**
-     * Calculate grade statistics
-     * @param {Array} attempts - Array of quiz attempts
-     * @returns {Object} Grade statistics
-     */
-    calculateGradeStatistics(attempts) {
-        if (!attempts || attempts.length === 0) {
-            return {
-                totalAttempts: 0,
-                averageScore: 0,
-                highestScore: 0,
-                lowestScore: 0,
-                passingRate: 0,
-                gradeDistribution: {}
-            };
-        }
-
-        const scores = attempts.map(attempt => attempt.score || 0);
-        const total = scores.reduce((sum, score) => sum + score, 0);
-        
-        // Calculate grade distribution
-        const gradeRanges = {
-            'A (90-100%)': scores.filter(s => s >= 90).length,
-            'B (80-89%)': scores.filter(s => s >= 80 && s < 90).length,
-            'C (70-79%)': scores.filter(s => s >= 70 && s < 80).length,
-            'D (60-69%)': scores.filter(s => s >= 60 && s < 70).length,
-            'F (Below 60%)': scores.filter(s => s < 60).length
-        };
-
-        return {
-            totalAttempts: attempts.length,
-            averageScore: Math.round((total / attempts.length) * 100) / 100,
-            highestScore: Math.max(...scores),
-            lowestScore: Math.min(...scores),
-            passingRate: Math.round((scores.filter(s => s >= 60).length / 60) * 100),
-            gradeDistribution: gradeRanges
-        };
-    }
-
-    /**
      * Cache for all results to avoid multiple API calls
      */
     _allResultsCache = null;
@@ -246,8 +151,8 @@ class ResultService {
         try {
             // Check cache first
             const now = Date.now();
-            if (this._allResultsCache && 
-                this._cacheTimestamp && 
+            if (this._allResultsCache &&
+                this._cacheTimestamp &&
                 (now - this._cacheTimestamp) < this._cacheExpiryMs) {
                 return this._allResultsCache;
             }
@@ -271,7 +176,7 @@ class ResultService {
             }
 
             const allResults = await response.json();
-            
+
             // Cache the results
             this._allResultsCache = allResults;
             this._cacheTimestamp = now;
@@ -286,7 +191,7 @@ class ResultService {
 
     /**
      * Get quiz results by category for a user (now uses cached batch call)
-     * @param {string} userId - User ID 
+     * @param {string} userId - User ID
      * @param {string} category - Quiz category
      * @returns {Promise<Object>} Quiz results for the category
      */
@@ -294,7 +199,7 @@ class ResultService {
         try {
             // Use the optimized batch call
             const allResults = await this.getAllResults(userId);
-            
+
             // Find the specific category (case insensitive)
             const categoryKey = Object.keys(allResults).find(
                 key => key.toLowerCase() === category.toLowerCase()
